'04-10'

+++ models/gamePlay.go
@@ -10,10 +10,10 @@ type GamePlay struct{
 	EndTime int64 `json:"end_time"`
 	StartTime int `json:"start_time"`
 	Type int `json:"type"`
-	Winmoney int `json:"winmoney"`
-	BMoney int `json:"b_money"`
-	Lostmoney int `json:"lostmoney"`
-	PtMoney int `json:"pt_money"`
+	Winmoney float64 `json:"winmoney"`
+	BMoney float64 `json:"b_money"`
+	Lostmoney float64 `json:"lostmoney"`
+	PtMoney float64 `json:"pt_money"`
 	IsQueue int `json:"is_queue"`
 	IsStatus int `json:"is_status"`
 }
